Implement output file header generation
Former-commit-id: a78afacb650a09daf032b5480357914f9e940e4d

+++ gti310/tp2/audio/AudioProperties.java
@@ -2,9 +2,9 @@ package gti310.tp2.audio;
 
 public class AudioProperties
 {
-	public int NumChannels;
+	public short NumChannels;
 	public int SampleRate;
-	public int BitsPerSample;
+	public short BitsPerSample;
 	
 	public int getFrameSize()
 	{
